Working on background
Working to see about adding some flair to the back ground image. Unfortunatley the box the image is in gets cut off when you add background color to the page. WIll need to work on that to see why and how to fix it. Maybe check out pg 244 in CSS book